backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
bcrypt==4.0.1
marshmallow==3.20.1
python-jose==3.3.0 